requirements.txt
pandas
python-dotenv
sqlalchemy
psycopg2
requests
google-cloud-storage
pandas
gcsfs
google-cloud-bigquery
cryptography